Added fully customizable dialog
Added the html structure for the consent dialog, which allows for
providing texts/labels aswell as layout in pure CSS, or by providing
the style in a javascript object.
In order to keep CSS an option, we provide only the absolute required
style, resulting in an ugly way too simple interface.

TODO: provide a configuration and a couple of basic styles